update services to latest rest api specs (#143)
* update services to latest rest api specs

* broken cosmos 2020-09

* wrong filter for operations

* only operations from listed input files

+++ services/mgmt/support/src/package_2020_04/models.rs
@@ -224,7 +224,7 @@ pub struct ServiceLevelAgreement {
     #[serde(rename = "expirationTime", skip_serializing)]
     pub expiration_time: Option<String>,
     #[serde(rename = "slaMinutes", skip_serializing)]
-    pub sla_minutes: Option<i64>,
+    pub sla_minutes: Option<i32>,
 }
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct SupportEngineer {
